Updated to allow documentation of other Sphinx-generated documentation as well as the Python standard library.
This was designed for generating docsets for NumPy and SciPy but should work with any Python modules that have been
documented using Sphinx.

The change has simply added processing of a single command-line argument, if present. If the script is called with no arguments
it behaves exactly as previously, generating documentation for whichever version of the Python standard library documentation
is in the current directory. If the script is called with one command-line argument, it simply sets the docset_name to the given
argument, and uses that whenever required in the code.

This allows the script to be called as ./create_docset.py numpy1.7 in a directory containing the NumPy HTML documentation
to produce a docset without any errors being generated.